revert some changes from 9822b7321
9822b7321dc1202209b503f3e06a51c4d9ac6214 accidentally included some
changes to ast which should not have been merged (possibly a bad
search-and-replace?)

This additionally un-exports mustParseVersion, which is only used within
the features package.
