+++ ast/version_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
-func TestMustParseVersion(t *testing.T) {
+func TestParseVersion(t *testing.T) {
 	namedReader := namedStringReader{strings.NewReader("VERSION 0.6")}
 	ver, err := ast.ParseVersionOpts(ast.FromReader(&namedReader))
 	r := require.New(t)
